Adding texture support. Color-enabled storage and rendering.
This time I struggled with the Display List leaks as I couldn't keep adding new features because of it. I also added texture support and color support (I haven't checked if alpha channel rendering is supported). The first components that will be used in the game have been added.

I have problems with the implementation of the game zone and the problem is that it contains 11,245,500 TILES!!!! if we consider the Tile structure now, then it takes 40 bytes.
That is, in the end it will all take ~ 428 MB. That's all for today, then I'll write another commit that no one will read.

Co-Authored-By: Fix <140541282+firf0x@users.noreply.github.com>