Adding new components and modifying existing components.
Today I decided to fix some problems that were bothering me, but because of them I had to make even more mistakes. For example (as for me, I think I incorrectly implemented the output of the rendering stream) outputting RenderThread into a separate stream and saving the link in the Tile struct to get the GameObject (which will cause boxing and unboxing).